my_requirements.txt
aiohttp==3.7.3
amqp==5.0.2
async-timeout==3.0.1
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==3.2.1
boto3==1.16.16
botocore==1.19.30
celery==5.0.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
dj-database-url==0.5.0
dj-pagination==2.5.0
Django==2.2.16
django-allauth==0.43.0
django-appconf==1.0.4
django-autocomplete-light==3.5.1
django-bootstrap3-datetimepicker-2==2.8.2
django-contrib-comments==1.9.2
django-cors-headers==3.5.0
django-downloadview==2.1.1
django-filter==2.4.0
django-floppyforms==1.9.0
django-forms-bootstrap==3.1.0
django-geonode-client==1.0.9
django-geonode-mapstore-client==2.0.8.1
django-grappelli==2.14.2
django-guardian==2.3.0
django-haystack==3.0
django-imagekit==4.0.2
django-invitations==1.9.3
django-ipware==3.0.2
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-jsonfield-compat==0.4.4
django-leaflet==0.27.1
django-mapstore-adapter==2.0.5.1
django-markdownify==0.8.1
django-modeltranslation==0.16.1
django-mptt==0.11.0
django-polymorphic==3.0.0
django-recaptcha==2.0.6
django-rest==0.8.6
django-rest-framework==0.1.0
django-storages==1.10.1
django-taggit==1.3.0
django-tastypie==0.14.3
django-tinymce==3.1.0
django-treebeard==4.3.1
django-uuid-upload-path==1.0.0
djangorestframework==3.12.2
djangorestframework-gis==0.16
djangorestframework-guardian==0.3.0
docker==4.4.0
drf-extensions==0.6.0
drf-writable-nested==0.6.2
drf-yasg==1.20.0
dynamic-rest==2.0.0
elasticsearch==7.10.0
factory-boy==3.1.0
Faker==5.0.0
flake8==3.8.4
flaky==3.7.0
future==0.18.2
geoip2==4.1.0
geolinks==0.2.0
# Editable install with no version control (GeoNode==3.1.0)
-e /home/peng/.virtualenvs/my_geonode/lib/python3.7/site-packages
geonode-announcements==2.0.2
geonode-avatar==5.0.7
geonode-dialogos==3.0.1
geonode-django-activity-stream==0.9.0
geonode-oauth-toolkit==2.1.0
geonode-user-messages==2.0.2
geoserver-restconfig==2.0.4.8
gisdata==0.5.4
glob2==0.7
gn-arcrest==10.5.3
gn-gsimporter==2.0.1
gunicorn==20.0.4
httplib2==0.18.1
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.1.1
incremental==17.5.0
inflection==0.3.1
iniconfig==1.1.1
invoke==1.4.1
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
jwcrypto==0.8
kombu==5.0.2
lxml==4.6.1
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
maxminddb==2.0.3
mccabe==0.6.1
mercantile==1.1.6
mock==4.0.2
multidict==5.1.0
numpy==1.19.4
oauthlib==3.1.0
OWSLib==0.20.0
packaging==20.7
parse==1.18.0
parse-type==0.5.2
parso==0.7.1
Paver==1.3.4
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==8.0.1
pinax-notifications==6.0.0
pinax-ratings==4.0.0
pluggy==0.13.1
prompt-toolkit==3.0.8
psutil==5.7.3
psycopg2==2.8.6
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pycountry==20.7.3
pycparser==2.20
pycsw==2.4.2
pyflakes==2.2.0
pygdal==2.4.2.6
Pygments==2.7.3
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==3.0.0.post1
pytest==6.1.2
pytest-bdd==4.0.1
pytest-django==4.1.0
pytest-splinter==3.3.0
python-dateutil==2.8.1
python-memcached==1.59
python-mimeparse==1.6.0
python-resize-image==1.1.19
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
selenium==3.141.0
Shapely==1.7.1
six==1.15.0
soupsieve==2.0.1
splinter==0.14.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
text-unidecode==1.3
timeout-decorator==0.5.0
toml==0.10.2
tqdm==4.53.0
traitlets==5.0.5
Twisted==20.3.0
typing-extensions==3.7.4.3
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.2
user-agents==2.2.0
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
wrapt==1.12.1
xmljson==0.2.1
xmltodict==0.12.0
yarl==1.6.3
zipp==3.4.0
zope.interface==5.2.0
